requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0